Widen the brightness factor from u8 to u16
Brightness was a u8 with 255 as unity, chosen so the divide collapsed
into a shift series - Cortex-M0/M0+ has no hardware divide. The cost was
resolution at the bottom of the range: the smallest non-zero multiplier
was 1/255, so WHITE at brightness 1 landed on channel 257 and at 2 on
514. Fading toward off banded visibly in the last few steps.

Unity is now 65535, matching the existing FULL constant and the
convention colors.rs already used. The shift-only property survives -
dividing by 65535 has the same reciprocal-series form - so no divide
helper is pulled into the scaling path.

Design notes:
- scale_channel takes a u16 factor and rounds to nearest. One correction
  term suffices over a 32-bit product; the widest intermediate is 32768
  short of u32::MAX, bounded by the operand types.
- div_65535 is its truncating sibling, replacing the three literal
  / 65535 divides in the HSV path. It reproduces them bit for bit, so
  hsv() output is unchanged and its tests needed no edits. The trailing
  + 1 is load-bearing: without it the series falls one short at every
  exact multiple of 65535.
- FULL_BRIGHTNESS is removed. At unity 65535 it duplicated FULL, which
  now serves as both the full-scale channel value and unity brightness.
- Q0.15 is untouched. Interpolation and easing keep ONE = 32768, which
  is what keeps a squared easing term inside a u32. Two scales, not one -
  README and CLAUDE.md now say so explicitly.

Tests go from 114 to 117. BRIGHTNESS_EPSILON drops from 256 to 8, since
its slack existed only to absorb u8 quantization. New coverage: the
bottom of the scale factor's range, the multiply path at the sequencer
level, and SequencerAction::SetBrightness dispatch, which had none.

Measured: flash 2100 -> 2104 B on thumbv6m and 1816 -> 1828 B on
thumbv7em. Struct sizes are unchanged for both the u32 and u64
instantiations - padding absorbs the wider field. service() on RP2040 is
0.33% slower, consistent across capacities; that is a stack-layout
codegen shift, not the unity comparison, which runs once per call and
would show a fixed absolute cost rather than a proportional one. RP2350
is unchanged within noise.

+++ src/color.rs
@@ -58,9 +58,11 @@ impl Rgb {
         }
     }
 
-    /// Scales every channel by an 8-bit factor, where 255 leaves the color unchanged.
+    /// Scales every channel by a 16-bit factor, where 65535 leaves the color unchanged.
+    ///
+    /// Note this is a different scale from [`lerp`](Self::lerp), which is Q0.15.
     #[inline]
-    pub const fn scale(self, factor: u8) -> Self {
+    pub const fn scale(self, factor: u16) -> Self {
         Self {
             r: scale_channel(self.r, factor),
             g: scale_channel(self.g, factor),
